refactor: PR리뷰 반영, import 경로 단축
