+++ apps/web/src/features/auth/model/mutations/useVerifyPasswordResetCodeMutation.ts
@@ -1,6 +1,6 @@
 import { useMutation, type UseMutationOptions } from '@tanstack/react-query';
 
-import { verifyPasswordResetCode } from '@/features/auth/api/post/session';
+import { verifyPasswordResetCode } from '@/features/auth/api';
 
 import type {
   PasswordResetVerifyRequestDto,
